Finish BasicTest.
I've finished the implementation of a Basic test to show how the flow of a device would work:

* The device is minted by the producer (an ERC721 id is created for that device) and the producer
becomes the current owner of the device.
* The device is rented to a consumer, who pays an amount of money to the producer (not implemented
yet), and receives the ownership of the device.
* The device could be subsequently passed from consumers to other consumers to recover a part of
the initial investment, respectively. Also a consumer can pass the device to a recycler.
* Once the recycler gets the device, it is destroyed (along with the ERC721 identifier) and the
recycler receives a part of the initial amount paid to the producer for the device.

Next steps would consist in improving the test by introducing the payments with respect to the
device and maybe adding more consumers.