Add new contract.
I've decided that the best way to implement the functionalities existing in the
recycle chain, is to include a new contract that manages both ERC721 and ERC20
instead of managing the interactions from the ERC721. This new contract is called
RecycleChain.

With this changes I had to update the test, which is not working yet.