Error in BasicTest.js
The error raises in line 67 of BasicTest.js, when we try to mint a device.
This error seems to be in line 106 on MyERC721.sol because if we remove
that line, seems to work. The error is something like "wrong return values.
Maybe run out of gas?". Some things to take into account:

* I'm using Transfer (for this method), so we don't need to use 'approve' method.
* During the migration we transfer 10.000 ERC20 tokens to ERC721 account
to handle the payments from that account. Previous approach to externalize
this into another contract was dismissed because of some issues passing the
initial msg.sender value among the different smart contracts.
* The event Transfer on ERC20 is emitted so it should work as expected, but
the error is raised.
* Before executing line 67 (which is a transaction), we run a call to the same
method in line 61 to get the the token_id of the new minted device.