Fix transferFrom issue.
Test now works until minting a new device. Next error appears on rent method,
when we try to transfer the ERC721 token to another owner, but it should
be easy to change in the smart contract. Also need to check the rest of the
methods that cause a transference with the ERC721 or ERC20 tokens.